Separate out web, parser tests
